requirements.txt
fastapi
uvicorn
pypdf
beautifulsoup4
requests
python-multipart


